test(api/note): add test suite for note service

+++ api/src/notes/notes.service.ts
@@ -6,23 +6,26 @@ import { NotesRepository } from 'src/notes/notes.repository';
 import { Result } from 'src/result';
 import { UsersService } from 'src/users/users.service';
 
-type CreateNoteError =
+type NoteError =
   | 'NOTE_CREATION_FAILED'
   | 'NOTE_RETRIEVAL_FAILED'
-  | 'USER_NOT_FOUND';
+  | 'USER_NOT_FOUND'
+  | 'NOTE_UPDATE_FAILED'
+  | 'NOTE_DELETION_FAILED'
+  | 'NOTE_ACCESS_FORBIDDEN'
+  | 'NOTE_NOT_FOUND';
 
 @Injectable()
 export class NotesService {
   constructor(
     private readonly notesRepository: NotesRepository,
     private readonly usersService: UsersService,
-    private readonly logger: Logger,
   ) {}
 
   async createNote(
     data: CreateNoteRequestDto,
     userId: number,
-  ): Promise<Result<Note, CreateNoteError>> {
+  ): Promise<Result<Note, NoteError>> {
     const user = await this.usersService.findOneById(userId);
 
     if (!user) {
@@ -30,8 +33,6 @@ export class NotesService {
     }
 
     try {
-      this.logger.debug(data);
-      this.logger.debug(user);
       const note = await this.notesRepository.create({
         ...data,
         author: {
@@ -45,7 +46,7 @@ export class NotesService {
     }
   }
 
-  async getAllNotes(userId: number): Promise<Result<Note[], CreateNoteError>> {
+  async getAllNotes(userId: number): Promise<Result<Note[], NoteError>> {
     const user = await this.usersService.findOneById(userId);
 
     if (!user) {
@@ -62,14 +63,50 @@ export class NotesService {
 
   async patchNote(
     noteId: number,
+    userId: number,
     data: PatchNoteRequestDto,
-  ): Promise<Result<Note, CreateNoteError>> {
-    const note = await this.notesRepository.patchNoteById(noteId, data);
-    return Result.ok(note);
+  ): Promise<Result<Note, NoteError>> {
+    const note = await this.notesRepository.getNoteById(noteId);
+
+    if (!note) {
+      return Result.err('NOTE_NOT_FOUND');
+    }
+
+    if (note.authorId !== userId) {
+      return Result.err('NOTE_ACCESS_FORBIDDEN');
+    }
+
+    try {
+      const updatedNote = await this.notesRepository.patchNoteById(
+        noteId,
+        data,
+      );
+      return Result.ok(updatedNote);
+    } catch (error) {
+      return Result.err('NOTE_UPDATE_FAILED');
+    }
   }
 
-  async deleteNote(noteId: number): Promise<Result<void, CreateNoteError>> {
-    await this.notesRepository.deleteNoteById(noteId);
-    return Result.ok(undefined);
+  async deleteNote(
+    noteId: number,
+    userId: number,
+  ): Promise<Result<void, NoteError>> {
+    const note = await this.notesRepository.getNoteById(noteId);
+
+    //no need to notice not found case. Consider it as ok
+    if (!note) {
+      return Result.ok(undefined);
+    }
+
+    if (note.authorId !== userId) {
+      return Result.err('NOTE_ACCESS_FORBIDDEN');
+    }
+
+    try {
+      await this.notesRepository.deleteNoteById(noteId);
+      return Result.ok(undefined);
+    } catch (error) {
+      return Result.err('NOTE_DELETION_FAILED');
+    }
   }
 }
